requirements.txt
flask>=3.0.0
pytest
pytest-playwright
pytest-html
requests
